# app/Models/User.php
<?php

namespace App\Models;

use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;
use Laravel\Passport\HasApiTokens;
use OpenApi\Annotations as OA;

/**
 * @OA\Schema(
 *  title="Usuário",
 *  @OA\Xml(
 *      name="Usuário"
 *  )
 * )
 */
class User extends Authenticatable
{
    use HasApiTokens, HasFactory, Notifiable;

    /**
     * The attributes that are mass assignable.
     *
     * @var string[]
     */
    protected $fillable = [
        'name',
        'email',
        'password',
    ];

    /**
     * The attributes that should be hidden for serialization.
     *
     * @var array
     */
    protected $hidden = [
        'password',
        'remember_token',
    ];

    /**
     * The attributes that should be cast.
     *
     * @var array
     */
    protected $casts = [
        'email_verified_at' => 'datetime',
    ];

    /**
     * @OA\Property(
     *  description = "ID do Usuário"
     * )
     * @var integer
     */
    private $id;

    /**
     * @OA\Property(
     *  description = "Nome do Usuário"
     * )
     * @var string
     */
    private $name;

    /**
     * @OA\Property(
     *  description = "E-mail do Usuário"
     * )
     * @var string
     */
    private $email;

    /**
     * @OA\Property(
     *  description = "Senha do Usuário"
     * )
     * @var string
     */
    private $password;
}